Removed most romaji
After testing some edge cases through various search queries on
jisho.org, I came to the conclusion that the removal of these romaji
shouldn't result in different behavior. If that is actually the case
(difficult for me to test) then I think they should be removed from the
database file, which I've done in this commit.